feat(*): replace demo-app with docs app
