+++ package.json
@@ -37,14 +37,14 @@
   },
   "dependencies": {
     "@angular/animations": "^18.0.3",
-    "@angular/cdk": "^18.0.3",
+    "@angular/cdk": "^18.0.5",
     "@angular/common": "^18.0.3",
     "@angular/compiler": "^18.0.3",
     "@angular/core": "^18.0.3",
-    "@angular/forms": "^18.0.3",
+    "@angular/forms": "^18.0.5",
     "@angular/platform-browser": "^18.0.3",
-    "@angular/platform-browser-dynamic": "^18.0.3",
-    "@angular/router": "^18.0.3",
+    "@angular/platform-browser-dynamic": "^18.0.5",
+    "@angular/router": "^18.0.5",
     "core-js": "^3.37.1",
     "moment": "^2.30.1",
     "moment-timezone": "^0.5.45",
@@ -53,12 +53,11 @@
     "zone.js": "^0.14.7"
   },
   "devDependencies": {
-    "@types/node": "^20.14.5",
-    "node-sass": "^9.0.0",
-    "@angular-devkit/build-angular": "^18.0.5",
-    "@angular/cli": "^18.0.5",
-    "@angular/compiler-cli": "^18.0.0",
+    "@angular-devkit/build-angular": "^18.0.6",
+    "@angular/cli": "^18.0.6",
+    "@angular/compiler-cli": "^18.0.5",
     "@types/jasmine": "~5.1.0",
+    "@types/node": "^20.14.9",
     "jasmine-core": "~5.1.2",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
@@ -66,6 +65,7 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
     "ng-packagr": "^18.0.0",
+    "node-sass": "^9.0.0",
     "typescript": "~5.4.5"
   }
 }
