--- projects/demo-app/src/main.ts
@@ -1,15 +1,0 @@
-import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
-import { bootstrapApplication } from '@angular/platform-browser';
-import { provideAnimations } from '@angular/platform-browser/animations';
-import { AppComponent } from './app/app.component';
-
-export const appConfig: ApplicationConfig = {
-  providers: [
-    provideZoneChangeDetection({ eventCoalescing: true }),
-    provideAnimations(),
-  ]
-};
-
-bootstrapApplication(AppComponent, appConfig).catch((err) => {
-  console.error(err)
-});
