--- projects/demo-app/src/styles.scss
@@ -1,7 +1,0 @@
-@import '../../picker/src/sass/picker.scss';
-
-html,
-body {
-  margin: 0;
-  background-color: #f0f0f0;
-}
