+++ projects/docs/public/favicon.ico
