+++ projects/docs/src/app/app.config.ts
@@ -1,0 +1,7 @@
+import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
+
+export const appConfig: ApplicationConfig = {
+  providers: [
+    provideZoneChangeDetection({ eventCoalescing: true })
+  ]
+};
